refactor: split host command history routes
