+++ src/backend/database/routes/host.ts
@@ -14,17 +14,7 @@ import {
   userRoles,
   sessionRecordings,
 } from "../db/schema.js";
-import {
-  eq,
-  and,
-  desc,
-  isNotNull,
-  or,
-  isNull,
-  gte,
-  sql,
-  inArray,
-} from "drizzle-orm";
+import { eq, and, isNotNull, or, isNull, gte, sql, inArray } from "drizzle-orm";
 import type { Request, Response } from "express";
 import axios from "axios";
 import multer from "multer";
@@ -47,6 +37,7 @@ import {
 import { registerHostOpksshRoutes } from "./host-opkssh-routes.js";
 import { registerHostFolderRoutes } from "./host-folder-routes.js";
 import { registerHostFileManagerBookmarkRoutes } from "./host-file-manager-bookmark-routes.js";
+import { registerHostCommandHistoryRoutes } from "./host-command-history-routes.js";
 
 const router = express.Router();
 
@@ -2046,141 +2037,7 @@ router.delete(
 
 registerHostFileManagerBookmarkRoutes(router, authenticateJWT);
 
-/**
- * @openapi
- * /host/command-history/{hostId}:
- *   get:
- *     summary: Get command history
- *     description: Retrieves the command history for a specific host.
- *     tags:
- *       - SSH
- *     parameters:
- *       - in: path
- *         name: hostId
- *         required: true
- *         schema:
- *           type: integer
- *     responses:
- *       200:
- *         description: A list of commands.
- *       400:
- *         description: Invalid userId or hostId.
- *       500:
- *         description: Failed to fetch command history.
- */
-router.get(
-  "/command-history/:hostId",
-  authenticateJWT,
-  async (req: Request, res: Response) => {
-    const userId = (req as AuthenticatedRequest).userId;
-    const hostIdParam = Array.isArray(req.params.hostId)
-      ? req.params.hostId[0]
-      : req.params.hostId;
-    const hostId = parseInt(hostIdParam, 10);
-
-    if (!isNonEmptyString(userId) || !hostId) {
-      sshLogger.warn("Invalid userId or hostId for command history fetch", {
-        operation: "command_history_fetch",
-        hostId,
-        userId,
-      });
-      return res.status(400).json({ error: "Invalid userId or hostId" });
-    }
-
-    try {
-      const history = await db
-        .select({
-          id: commandHistory.id,
-          command: commandHistory.command,
-        })
-        .from(commandHistory)
-        .where(
-          and(
-            eq(commandHistory.userId, userId),
-            eq(commandHistory.hostId, hostId),
-          ),
-        )
-        .orderBy(desc(commandHistory.executedAt))
-        .limit(200);
-
-      res.json(history.map((h) => h.command));
-    } catch (err) {
-      sshLogger.error("Failed to fetch command history from database", err, {
-        operation: "command_history_fetch",
-        hostId,
-        userId,
-      });
-      res.status(500).json({ error: "Failed to fetch command history" });
-    }
-  },
-);
-
-/**
- * @openapi
- * /host/command-history:
- *   delete:
- *     summary: Delete command from history
- *     description: Deletes a specific command from the history of a host.
- *     tags:
- *       - SSH
- *     requestBody:
- *       required: true
- *       content:
- *         application/json:
- *           schema:
- *             type: object
- *             properties:
- *               hostId:
- *                 type: integer
- *               command:
- *                 type: string
- *     responses:
- *       200:
- *         description: Command deleted from history.
- *       400:
- *         description: Invalid data.
- *       500:
- *         description: Failed to delete command.
- */
-router.delete(
-  "/command-history",
-  authenticateJWT,
-  async (req: Request, res: Response) => {
-    const userId = (req as AuthenticatedRequest).userId;
-    const { hostId, command } = req.body;
-
-    if (!isNonEmptyString(userId) || !hostId || !command) {
-      sshLogger.warn("Invalid data for command history deletion", {
-        operation: "command_history_delete",
-        hostId,
-        userId,
-      });
-      return res.status(400).json({ error: "Invalid data" });
-    }
-
-    try {
-      await db
-        .delete(commandHistory)
-        .where(
-          and(
-            eq(commandHistory.userId, userId),
-            eq(commandHistory.hostId, hostId),
-            eq(commandHistory.command, command),
-          ),
-        );
-
-      res.json({ message: "Command deleted from history" });
-    } catch (err) {
-      sshLogger.error("Failed to delete command from history", err, {
-        operation: "command_history_delete",
-        hostId,
-        userId,
-        command,
-      });
-      res.status(500).json({ error: "Failed to delete command" });
-    }
-  },
-);
+registerHostCommandHistoryRoutes(router, authenticateJWT);
 
 async function resolveHostCredentials(
   host: Record<string, unknown>,
